requirements.txt
simplejson
zeep>=3.2.0
python-stdnum>=1.17
pycountry>=19.8.18
